# tests/unit_tests/fixtures/openapi.yaml
openapi: 3.0.0
info:
  title: V0 Semantic layer API
  version: 1.0.11
  contact:
    name: Андрей Абакин
    url: https://addressbook/search/profile/1763632
    email: abakin.a.v@sberbank.ru
  description: Semantic layer API for EPMPlatform
servers:
  - url: https://ci00163863-epmp-sndbx.apps.dev-terra000004-idm.ocp.delta.sbrf.ru
    description: DEV стенд
  - url: https://ci04763767-semlayer.apps.ift-terra000039-edb.ocp.delta.sbrf.ru
    description: IFT стенд
paths:
...
  /api/v0/tenants/{tenantName}/models/{modelName}/measures/:
    get:
      parameters:
        - example: 1
          name: page
          description: Страница (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationPageCount"
          in: query
          required: false
        - example: 20
          name: pageSize
          description: Количество строк (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationRowCount"
          in: query
          required: false
        - name: sortDirection
          description: Направление сортировки
          schema:
            $ref: "#/components/schemas/sortDirection"
          in: query
          required: false
      tags:
        - v0
      summary: Get all measures by model name
      operationId: getMeasuresByModelName
      responses:
        "200":
          description: Все показатели модели `modelName`
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/measure"
                maxItems: 1000
                uniqueItems: true
        "500":
          $ref: "#/components/responses/unexpectedError"
    post:
      tags:
        - v0
      summary: Get measures by model name and measure name
      operationId: getMeasuresByModelNameAndMeasureNames
      responses:
        "201":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/measure"
                maxItems: 1000
                uniqueItems: true
          description: Показатели модели `modelName` по списку `measureNames`
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Показатели не найдены в модели `modelName`
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/measureName"
              maxItems: 1000
              uniqueItems: true
        description: New measure
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/measures/copyToAnotherModel:
    post:
      tags:
        - v0
      summary: Copy measure to another model
      operationId: copyMeasureToAnotherModel
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/respCopyObject"
          description: Измерение создано в указанных моделях
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
          description: Измерение уже существует в модел `modelName`
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/requestCopyObject"
      description: New model for measure
    parameters:
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/models/{modelName}/measures/create:
    post:
      tags:
        - v0
      summary: Create measure by model name
      operationId: createMeasureByModelName
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/measure"
          description: Показатель создан в модели `modelName`
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Показатель уже существует в модели `modelName`
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/measure"
        description: New measure
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/models/{modelName}/measures/{measureName}:
    get:
      tags:
        - v0
      summary: Get measure by model name and measure name
      operationId: getMeasureByModelNameAndMeasureName
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/measure"
          description: Показатель `measureName` модели `modelName`
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Не найден показатель `measureName` модели `modelName`
        "500":
          $ref: "#/components/responses/unexpectedError"
    delete:
      tags:
        - v0
      summary: Delete measure by model name and measure name
      operationId: deleteMeasureByModelNameAndMeasureName
      responses:
        "200":
          description: Показатель `measureName` модели `modelName` удален
        "404":
          description: Показатель `measureName` модели `modelName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
    patch:
      tags:
        - v0
      summary: Change measure by model name and measure name
      operationId: changeMeasureByModelNameAndMeasureName
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/measure"
          description: Показатель `measureName` модели `modelName` изменен
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Показатель `measureName` модели `modelName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/measureEdit"
    parameters:
      - $ref: "#/components/parameters/measureName"
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
...
components:
  responses:
    unexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
  parameters:
    modelName:
      name: modelName
      description: Уникальный идентификатор модели
      schema:
        $ref: "#/components/schemas/modelName"
      in: path
      required: true
...
    measureName:
      name: measureName
      description: Уникальный идентификатор показателя
      schema:
        $ref: "#/components/schemas/measureName"
      in: path
      required: true
...
    tenantName:
      name: tenantName
      description: Название тенанта
      schema:
        $ref: "#/components/schemas/tenantName"
      in: path
      required: true
...
  schemas:
    tenantName:
      $ref: "#/components/schemas/objectName32"
    dimensionName:
      $ref: "#/components/schemas/objectName32"
...
    measureName:
      $ref: "#/components/schemas/objectName32"
...
    modelName:
      $ref: "#/components/schemas/objectName32"
    objectName32:
      description: Уникальное техническое имя объекта
      type: string
      minLength: 1
      maxLength: 32
      pattern: "^[a-z0-9_]+$"
      example: "objectname32"
...
    requestCopyObject:
      type: object
      additionalProperties: false
      properties:
        modelNames:
          type: array
          items:
            $ref: "#/components/schemas/modelName"
          maxItems: 1000
          uniqueItems: true
        objects:
          type: array
          items:
            $ref: "#/components/schemas/objectName32"
          maxItems: 1000
          uniqueItems: true
...
    respCopyObject:
      type: array
      maxItems: 1000
      uniqueItems: true
      items:
        type: object
        additionalProperties: false
        properties:
          tenant:
            $ref: "#/components/schemas/tenantName"
          objectName:
            $ref: "#/components/schemas/objectName32"
          modelNames:
            type: array
            items:
              $ref: "#/components/schemas/modelName"
            maxItems: 1000
            uniqueItems: true
          result:
            $ref: "#/components/schemas/result"
          comment:
            pattern: '^[^\n\t]+$'
            type: string
            maxLength: 255
            description: Комментарий
...
    httpValidationError:
      type: object
      additionalProperties: false
      properties:
        detail:
          type: array
          maxItems: 1000
          uniqueItems: true
          items:
            type: object
            additionalProperties: false
            properties:
              loc:
                type: array
                items:
                  type: string
                  maxLength: 255
                  pattern: '^[^\n\t]+$'
                maxItems: 1000
                uniqueItems: true
              msg:
                $ref: "#/components/schemas/msg"
              type:
                type: string
                maxLength: 255
                pattern: '^[^\n\t]+$'
                description: Тип ошибки
...
    error:
      required:
        - message
      type: object
      additionalProperties: false
      properties:
        message:
          description: A human readable error message
          type: string
          pattern: "^[a-zA-Z0-9_]+$"
          minLength: 1
          maxLength: 255
    measure:
      allOf:
        - $ref: "#/components/schemas/versionInfo"
        - $ref: "#/components/schemas/measureEdit"
        - required:
            - name
          type: object
          additionalProperties: false
          properties:
            name:
              $ref: "#/components/schemas/measureName"
            aorType:
              $ref: '#/components/schemas/aorType'
            models:
              type: array
              items:
                $ref: "#/components/schemas/modelName"
              maxItems: 1000
              minItems: 0
              uniqueItems: true
              description: Массив имен моделей, к которым относится показатель, со статусом обработки объекта в ней
    dimensionFilter:
      type: object
      additionalProperties: false
      required:
        - dimensionName
      properties:
        dimensionName:
          $ref: "#/components/schemas/dimensionName"
        dimensionValue:
          description: Значение dimension
          type: string
          maxLength: 255
          pattern: "^[a-zA-Z0-9_]+$"
...
    objectLabels:
      type: array
      maxItems: 100
      uniqueItems: true
      description: |
        Тексты, которые описывают различные сущности системы на разных языках для
          - признаков
          - атрибутов признаков
          - показателей
          - хранилищ данных
          - композитов
          - ...
      items:
        type: object
        additionalProperties: false
        required:
          - language
          - type
        description: Зависимый от языка текст различной длины
        properties:
          language:
            type: string
            maxLength: 5
            description: Язык текста
            example: ru-ru
            pattern: "^[a-z]{2}-[a-z]{2}$"
          type:
            $ref: "#/components/schemas/labelType"
          text:
            type: string
            description: Отображаемый текст
            pattern: '^[^\n\t]+$'
            minLength: 1
            maxLength: 255
...
    aggregationType:
      description: Тип агрегации значений показателя или поля
      maxLength: 20
      enum:
        - SUMMATION
        - MAXIMUM
        - MINIMUM
        - NO AGGREGATION
      type: string
...
    measureKeyType:
      description: Технический тип данных поля для показателя
      enum:
        - FLOAT
        - INTEGER
        - DECIMAL
      type: string
      maxLength: 20
    labelType:
      description: Вид метки для описания объекта - короткий для вывода в поле ограниченной длины, - длинный для отображения подробного описания
      type: string
      maxLength: 20
      enum:
        - SHORT
        - LONG
...
    measureEdit:
      type: object
      additionalProperties: false
      properties:
        labels:
          $ref: "#/components/schemas/objectLabels"
        measureRef:
          $ref: "#/components/schemas/measureName"
        type:
          $ref: "#/components/schemas/measureKeyType"
        precision:
          description: При изменении только в сторону увеличения
          type: integer
          minimum: 1
          maximum: 34
          default: 0
        scale:
          description: При изменении только в сторону увеличения, актуально для типа данных Decimal
          type: integer
          minimum: 0
          maximum: 34
        authorizationRelevant:
          description: authorization relevance
          type: boolean
          default: false
        aggregationType:
          $ref: "#/components/schemas/aggregationType"
        unitOfMeasure:
          oneOf:
            - $ref: "#/components/schemas/dimensionName"
            - $ref: "#/components/schemas/dimensionFilter"
        filter:
          type: array
          description: Список уникальных значений dimension
          items:
            $ref: "#/components/schemas/dimensionFilter"
          maxItems: 100
          uniqueItems: true
...
    sortDirection:
      description: Направление сортировки
      default: asc
      enum:
        - asc
        - desc
      type: string
    paginationPageCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
    paginationRowCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
...
    result:
      description: Результат
      type: string
      enum:
        - SUCCESS
        - PENDING
        - FAILURE
    msg:
      type: string
      maxLength: 8192
      minLength: 1
      description: Сообщение
      example: Комментарий
...
    aorType:
      description: Тип объекта в AOR
      type: string
      enum:
        - sl_model
        - sl_hierarchy
        - sl_dimension
        - sl_measure
        - sl_datastorage
        - sl_databaseobject
        - sl_composite
        - sl_database
        - sl_aigroup
      example: sl_dimension
...
    version:
      description: Версия объекта
      type: integer
      minimum: 1
      maximum: 2147483647
      example: 6
...
    timestamp:
      type: string
      description: Дата и время
      example: "2025-05-06 18:16:15.690"
    user:
      type: string
      description: Пользователь
      pattern: ^[^\n\t]+$
      example: EPMSuperUser
...
    versionInfo:
      description: Все версии объекта без имени
      type: object
      properties:
        version:
          $ref: "#/components/schemas/version"
        updatedAt:
          $ref: "#/components/schemas/timestamp"
        updatedBy:
          $ref: "#/components/schemas/user"